move NestedProperty
